+++ backend/src/main/java/carteirafiis/carteira/service/FiiService.java
@@ -2,7 +2,9 @@ package carteirafiis.carteira.service;
 
 
 import carteirafiis.carteira.model.FiiModel;
+import carteirafiis.carteira.model.UserModel;
 import carteirafiis.carteira.repository.FiiRepository;
+import carteirafiis.carteira.security.AuthUtil;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -11,9 +13,11 @@ import org.springframework.stereotype.Service;
 public class FiiService {
 
     private final FiiRepository fiiRepository;
+    private final AuthUtil authUtil;
 
-    public FiiService(FiiRepository fiiRepository) {
+    public FiiService(FiiRepository fiiRepository, AuthUtil authUtil) {
         this.fiiRepository = fiiRepository;
+        this.authUtil = authUtil;
     }
 
     // create fii (post)
@@ -25,12 +29,19 @@ public class FiiService {
     // listar fiis (get)
 
     public Page<FiiModel> listFii(Pageable pageable){
-        return fiiRepository.findAll(pageable);
+        UserModel user = authUtil.getLoggedUser();
+        return fiiRepository.findByUserId(user.getId(), pageable);
     }
 
     // atualizar fii (put)
 
     public void updateFii(FiiModel fiiModel){
+        UserModel user = authUtil.getLoggedUser();
+
+        if (!fiiModel.getUser().getId().equals(user.getId())) {
+            throw new RuntimeException("you do not have permission to update this FII");
+        }
+
         fiiRepository.save(fiiModel);
     }
 
@@ -41,6 +52,12 @@ public class FiiService {
 
     // delete fii (delete)
     public void deleteFii(FiiModel fiiModel){
+        UserModel user = authUtil.getLoggedUser();
+
+        if(!fiiModel.getUser().getId().equals(user.getId())){
+            throw new RuntimeException("you do not have permission to delete this FII");
+        }
+
         fiiRepository.deleteById(fiiModel.getId());
     }
